requirements.txt
Flask>=3.0.3,<3.2
flask-cors==4.0.0
python-dotenv==1.0.1
gunicorn==21.2.0
PyJWT==2.8.0
Werkzeug==3.0.4

